@@ -1,12 +1,12 @@
 {
-  "name": "@brisanet/ion",
-  "version": "0.0.117",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/Brisanet/ion.git"
-  },
+  "name": "ion",
+  "version": "0.0.1",
   "peerDependencies": {
-    "@angular/common": "^8.1.3",
-    "@angular/core": "^8.1.3"
-  }
+    "@angular/common": "^19.2.0",
+    "@angular/core": "^19.2.0"
+  },
+  "dependencies": {
+    "tslib": "^2.3.0"
+  },
+  "sideEffects": false
 }
