@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { IonComponent } from './ion.component';
 
@@ -6,13 +6,12 @@ describe('IonComponent', () => {
   let component: IonComponent;
   let fixture: ComponentFixture<IonComponent>;
 
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [IonComponent],
-    }).compileComponents();
-  }));
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [IonComponent]
+    })
+    .compileComponents();
 
-  beforeEach(() => {
     fixture = TestBed.createComponent(IonComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
