@@ -2,7 +2,14 @@ import { Component } from '@angular/core';
 
 @Component({
   selector: 'lib-ion',
-  template: ` <p>ion works!</p> `,
-  styles: [],
+  imports: [],
+  template: `
+    <p>
+      ion works!
+    </p>
+  `,
+  styles: ``
 })
-export class IonComponent {}
+export class IonComponent {
+
+}
