@@ -3,10 +3,14 @@ import { TestBed } from '@angular/core/testing';
 import { IonService } from './ion.service';
 
 describe('IonService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}));
+  let service: IonService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(IonService);
+  });
 
   it('should be created', () => {
-    const service: IonService = TestBed.get(IonService);
     expect(service).toBeTruthy();
   });
 });
