@@ -1,6 +1,9 @@
 import { Injectable } from '@angular/core';
 
 @Injectable({
-  providedIn: 'root',
+  providedIn: 'root'
 })
-export class IonService {}
+export class IonService {
+
+  constructor() { }
+}
