@@ -1,6 +1,6 @@
 export * from './createElement';
-export * from './datePicker';
 export * from './dropdown';
 export * from './generateID';
 export * from './setTimer';
 export * from './validateHexColor';
+export * from './pipes';
