@@ -3,9 +3,10 @@ import { SafeAny } from '../../safe-any';
 
 @Pipe({
   name: 'replaceEmpty',
+  standalone: true,
 })
 export class ReplaceEmptyPipe implements PipeTransform {
-  transform(value: SafeAny, replaceValue: string): string {
+  transform(value: SafeAny, replaceValue: string): SafeAny {
     if (value === null || value === undefined || value === '') {
       return replaceValue;
     }
