@@ -1,63 +1,3 @@
-/*
- * Public API Surface of ion
- */
-export { default as BnTable } from './core/bn-table/bn-table';
-export * from './lib/accordion/accordion.module';
-export * from './lib/alert/alert.module';
-export * from './lib/avatar/avatar.module';
-export * from './lib/badge/badge.module';
-export * from './lib/breadcrumb/breadcrumb.module';
-export * from './lib/button/button.module';
-export * from './lib/card/card.module';
-export * from './lib/checkbox/checkbox.module';
-export * from './lib/chip-group/chip-group.module';
-export * from './lib/chip/chip.module';
-export * from './lib/col/col.module';
-export * from './lib/core/types';
-export * from './lib/divider/divider.module';
-export * from './lib/dropdown/dropdown.module';
-export * from './lib/icon/icon.module';
-export * from './lib/indicator/indicator.module';
-export * from './lib/info-badge/info-badge.module';
-export * from './lib/input-area/input-area.module';
-export * from './lib/input-counter/input-counter.module';
-export * from './lib/input-select/input-select.module';
-export * from './lib/input/input.module';
-export * from './lib/ion.component';
-export * from './lib/ion.module';
 export * from './lib/ion.service';
-export * from './lib/message/message.module';
-export * from './lib/modal/modal.module';
-export * from './lib/modal/modal.service';
-export * from './lib/no-data/no-data.module';
-export * from './lib/notification/notification.module';
-export * from './lib/notification/service/notification.service';
-export * from './lib/pagination/pagination.module';
-export * from './lib/picker/date-picker/date-picker.module';
-export * from './lib/popconfirm/popconfirm.module';
-export * from './lib/popover/popover.module';
-export * from './lib/position/position.service';
-export * from './lib/radio-group/radio-group.module';
-export * from './lib/radio/radio.module';
-export * from './lib/row/row.module';
-export * from './lib/select/select.module';
-export * from './lib/shared.module';
-export * from './lib/sidebar/sidebar.module';
-export * from './lib/simple-menu/simple-menu.module';
-export * from './lib/skeleton/skeleton.module';
-export * from './lib/smart-table/smart-table.module';
-export * from './lib/spinner/spinner.module';
-export * from './lib/step/step.module';
-export * from './lib/switch/switch.module';
-export * from './lib/tab-group/tab-group.module';
-export * from './lib/tab/tab.module';
-export * from './lib/table/table.module';
-export * from './lib/table/utilsTable';
-export * from './lib/tag/tag.module';
-export * from './lib/theme';
-export * from './lib/tooltip/tooltip.module';
-export * from './lib/tour';
-export * from './lib/triple-toggle/triple-toggle.module';
-export * from './lib/typography/';
-export { default as debounce } from './lib/utils/debounce';
-export * from './lib/link/link.module';
+export * from './lib/ion.component';
+export * from './lib/button/button.component';
